+++ pkg/local_object_storage/shard/delete.go
@@ -81,9 +81,16 @@ func (s *Shard) delete(prm DeletePrm) (DeleteRes, error) {
 	s.decObjectCounterBy(physical, res.RawObjectsRemoved())
 	s.decObjectCounterBy(logical, res.AvailableObjectsRemoved())
 
+	var totalRemovedPayload uint64
+	removedSizes := res.RemovedObjectSizes()
+
 	for i := range prm.addr {
-		s.addToContainerSize(prm.addr[i].Container().EncodeToString(), -int64(res.RemovedObjectSizes()[i]))
+		removedPayload := removedSizes[i]
+		totalRemovedPayload += removedPayload
+		s.addToContainerSize(prm.addr[i].Container().EncodeToString(), -int64(removedPayload))
 	}
+
+	s.addToPayloadCounter(-int64(totalRemovedPayload))
 
 	for i := range prm.addr {
 		var delPrm common.DeletePrm