+++ pkg/local_object_storage/shard/shard.go
@@ -56,6 +56,9 @@ type MetricsWriter interface {
 	// AddToContainerSize must add a value to the container size.
 	// Value can be negative.
 	AddToContainerSize(cnr string, value int64)
+	// AddToPayloadSize must add a value to the payload size.
+	// Value can be negative.
+	AddToPayloadSize(value int64)
 	// IncObjectCounter must increment shard's object counter taking into account
 	// object type.
 	IncObjectCounter(objectType string)
@@ -346,6 +349,8 @@ func (s *Shard) initMetrics() {
 			return
 		}
 
+		var totalPayload uint64
+
 		for i := range cnrList {
 			size, err := s.metaBase.ContainerSize(cnrList[i])
 			if err != nil {
@@ -355,7 +360,10 @@ func (s *Shard) initMetrics() {
 				continue
 			}
 			s.metricsWriter.AddToContainerSize(cnrList[i].EncodeToString(), int64(size))
+			totalPayload += size
 		}
+
+		s.metricsWriter.AddToPayloadSize(int64(totalPayload))
 	}
 }
 
@@ -377,5 +385,11 @@ func (s *Shard) decObjectCounterBy(typ string, v uint64) {
 func (s *Shard) addToContainerSize(cnr string, size int64) {
 	if s.cfg.metricsWriter != nil {
 		s.cfg.metricsWriter.AddToContainerSize(cnr, size)
+	}
+}
+
+func (s *Shard) addToPayloadCounter(size int64) {
+	if s.cfg.metricsWriter != nil {
+		s.cfg.metricsWriter.AddToPayloadSize(size)
 	}
 }